+++ jdaviz/configs/default/plugins/plot_options/plot_options.py
@@ -19,6 +19,7 @@ from jdaviz.core.template_mixin import (PluginTemplateMixin, ViewerSelect, Layer
 from jdaviz.core.user_api import PluginUserApi
 from jdaviz.core.tools import ICON_DIR
 from jdaviz.utils import bqplot_clear_figure
+from jdaviz.core.marks import HistogramMark
 
 __all__ = ['PlotOptions']
 
@@ -609,3 +610,31 @@ class PlotOptions(PluginTemplateMixin):
         if len(sub_data) > 0:
             hist_lims = interval.get_limits(sub_data)
             hist_mark.min, hist_mark.max = hist_lims
+
+        self._check_if_v_stretch_changed()
+
+    @observe('stretch_vmin_value', 'stretch_vmax_value')
+    def _check_if_v_stretch_changed(self, msg=None):
+        self._remove_histogram_marks()
+        self._add_histogram_marks()
+
+    def _add_histogram_marks(self):
+        if self.stretch_histogram is None:
+            return
+        scales = {'x': self.stretch_histogram.axes[0].scale, 'y': bqplot.LinearScale()}
+        v_stretch_lines = []
+        if self.stretch_vmin.value >= self.stretch_histogram.marks[0].min:
+            vmin_lines = HistogramMark(min_max_value=[self.stretch_vmin.value, self.stretch_vmin.value], # noqa
+                                       scales=scales)
+            v_stretch_lines.append(vmin_lines)
+        if self.stretch_vmax.value <= self.stretch_histogram.marks[0].max:
+            vmax_lines = HistogramMark(min_max_value=[self.stretch_vmax.value, self.stretch_vmax.value], # noqa
+                                       scales=scales)
+            v_stretch_lines.append(vmax_lines)
+        self.stretch_histogram.marks = self.stretch_histogram.marks + v_stretch_lines
+
+    def _remove_histogram_marks(self):
+        if self.stretch_histogram is None:
+            return
+        self.stretch_histogram.marks = [mark for mark in self.stretch_histogram.marks
+                                        if not isinstance(mark, HistogramMark)]