+++ jdaviz/core/marks.py
@@ -631,3 +631,13 @@ class MarkersMark(PluginScatter):
     def __init__(self, viewer, **kwargs):
         kwargs.setdefault('marker', 'circle')
         super().__init__(viewer, **kwargs)
+
+
+class HistogramMark(Lines):
+    def __init__(self, min_max_value, scales, **kwargs):
+        # Vertical line in LinearScale
+        y = [0, 1]
+        colors = ["#c75d2c"]
+        line_style = "solid"
+        super().__init__(x=min_max_value, y=y, scales=scales, colors=colors, line_style=line_style,
+                         **kwargs)