+++ CHANGES.rst
@@ -22,6 +22,8 @@ New Features
 
 - Line Analysis "Redshift from Centroid" only visible when lines are loaded. [#2294]
 
+- Add lines representing the stretch vmin and vmax to the plot options histogram. [#2301]
+
 Cubeviz
 ^^^^^^^